COMP: Fixed warnings (made resolution an int instead of a double)

+++ Qt/Widgets/pqDoubleRangeWidget.cxx
@@ -75,13 +75,13 @@ pqDoubleRangeWidget::~pqDoubleRangeWidget()
 }
 
 //-----------------------------------------------------------------------------
-double pqDoubleRangeWidget::resolution() const
+int pqDoubleRangeWidget::resolution() const
 {
   return this->Resolution;
 }
 
 //-----------------------------------------------------------------------------
-void pqDoubleRangeWidget::setResolution(double val)
+void pqDoubleRangeWidget::setResolution(int val)
 {
   this->Resolution = val;
   this->Slider->setRange(0, this->Resolution);
@@ -179,7 +179,7 @@ void pqDoubleRangeWidget::sliderChanged(int val)
 {
   if(!this->BlockUpdate)
     {
-    double fraction = val / this->Resolution;
+    double fraction = val / static_cast<double>(this->Resolution);
     double range = this->Maximum - this->Minimum;
     double v = (fraction * range) + this->Minimum;
     this->BlockUpdate = true;
@@ -199,7 +199,7 @@ void pqDoubleRangeWidget::textChanged(const QString& text)
     this->BlockUpdate = true;
     double range = this->Maximum - this->Minimum;
     double fraction = (val - this->Minimum) / range;
-    int sliderVal = qRound(fraction * this->Resolution);
+    int sliderVal = qRound(fraction * static_cast<double>(this->Resolution));
     this->Slider->setValue(sliderVal);
     this->setValue(val);
     this->BlockUpdate = false;
@@ -218,7 +218,7 @@ void pqDoubleRangeWidget::updateSlider()
   this->Slider->blockSignals(true);
   double range = this->Maximum - this->Minimum;
   double fraction = (this->Value - this->Minimum) / range;
-  int v = qRound(fraction * this->Resolution);
+  int v = qRound(fraction * static_cast<double>(this->Resolution));
   this->Slider->setValue(v);
   this->Slider->blockSignals(false);
 }
